order-service/src/main/java/com/qthuy2k1/orderservice/controller/OrderController.java
package com.qthuy2k1.orderservice.controller;

import com.qthuy2k1.orderservice.dto.request.OrderRequest;
import com.qthuy2k1.orderservice.dto.response.ApiResponse;
import com.qthuy2k1.orderservice.dto.response.OrderResponse;
import com.qthuy2k1.orderservice.service.OrderService;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Positive;
import lombok.AccessLevel;
import lombok.RequiredArgsConstructor;
import lombok.experimental.FieldDefaults;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequiredArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@RequestMapping("/orders")
public class OrderController {
    OrderService orderService;

    @PostMapping
    public ResponseEntity<ApiResponse<OrderResponse>> createOrder(@RequestBody @Valid OrderRequest orderRequest) {
        return ResponseEntity.status(HttpStatus.CREATED).body(
                ApiResponse.<OrderResponse>builder()
                        .result(orderService.createOrder(orderRequest))
                        .build()
        );
    }

    @PutMapping("{id}")
    public ResponseEntity<String> updateOrder(@PathVariable("id") @Positive int id, @RequestBody OrderRequest orderRequest) {
        orderService.updateOrder(id, orderRequest);
        return new ResponseEntity<>("Success", HttpStatus.OK);
    }
}
